requirements-api.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.9
google-genai>=1.52.0
httpx>=0.28.1
Pillow==10.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.32.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
stripe==7.9.0
pydantic[email]==2.9.2
boto3==1.35.96